+++ src/Smartstore.Web/Models/ShoppingCart/Mappers/CartItemQuantityInputMapper.cs
@@ -45,7 +45,7 @@ namespace Smartstore.Web.Models.Cart
             model.MaxOrderAmount = product.OrderMaximumQuantity;
             model.QuantityStep = product.QuantityStep;
 
-            if (product.ManageInventoryMethod == ManageInventoryMethod.ManageStock || product.ManageInventoryMethod == ManageInventoryMethod.ManageStockByAttributes)
+            if (product.ManageInventoryMethod > ManageInventoryMethod.DontManageStock)
             {
                 model.MaxInStock = product.StockQuantity;
             }